Fix event tracking when page is translated
When Google Chrome's built-in Google Translate extension
translates pages, it modifies not just the text but also the
HTML structure of a document.

Before translation:

```
<a href="/standard-visitor-visa">Standard Visitor visa</a>
```

After translation:

```
<a href="/standard-visitor-visa"><font style="vertical-align: inherit;"><font style="vertical-align: inherit;">demande de visa de visiteur standard</font></font></a>
```

The Event Listener is attached to the `font` element, rather than
the `a` element, which is a problem as the `font` doesn't have the
href (what we supply as the event action in the Google Analytics call).

The solution is to use `currentTarget` rather than `target`, as this
refers to the element the event handler is attached to, rather than
the potentially descendent element on which the event occurred.

https://developer.mozilla.org/en-US/docs/Web/API/Event/currentTarget

From some manual testing this seems to resolve the problem.