Use smart answer generator to create initial setup
Use the smart-answer generator to create a new smart-answer to help
users find the travel rules that pertain to their trip.

The url is a placeholder as the final url will match the title of the
smart-answer, but neither have been determined yet.

It's also common for the url to also match the class names for the
smart-answer, so some refactoring, will need to be done later to keep
the names in sync.