Add a dirty hack to memoize `self.calculator`
We can only set `self.calculator` in a `node` scope, not a `flow` scope.
This is why, traditionally, the `calculator` object is set in the
`on_response` of the first question as this is the first time we access
a `node` scope.

In this flow however, the first question is in a loop. But we only want
to create the calculator object once.

Unfortunately we can't use normal memoization like

```
self.calculator ||= SmartAnswer::Calculators::CovidTravelAbroadCalculator.new
```

As it falls foul of the [method missing] exception handler.

This dirty hack is acceptable for this spike, but something better would
be need to be implemented before going to production. For example,
revisiting the need of the method missing check. Or creating an
initialize block for flows to allow calculator objects to be created
before the first question.

[method missing]: https://github.com/alphagov/smart-answers/blob/fb46663d23c6236e34afb089c6ea1bca59653b3f/lib/smart_answer/state.rb#L12-L22