Add dynamic checkbox question
We had to change the flow to be a query parameters because the options method doesn't work as expected with URL based questions. This was in order to get the dynamic checkboxes working. This meant that we had to remove the question marks at the end of question names. Question marks cause it to behave funnily and loop where it shouldn't.