Outcome changes
I made some changes to the outcome view as some of the logic in there
did not seem necessary. Specifically there seems to be AND or OR
statements that seem to be a given or are impossible to reach in the
you_can_use_result view.

The home costs conditional is a good example of this, previously the
code had a conditional requiring this to have a value above 0. This is
irrelevant in this view as any home_costs value below 25 would lead to a
different outcome view entirely.