Add questions to get any other countries
Add questions to ask users if they want to add any other countries and
collect the countries they are travelling to.

These questions have been added together as they call each other in a
loop.

A global variable had to be created at the class level to control the
loop as it's not possible to access a calculator object outside of
certain scopes (e.g. `on_response`, `next_node`). So rather than asking
users how many countries they are travelling to and then using that
count to loop on the "which country" question, another question had be
added to allow users to break the loop.