Add travel rules for each world destination
These are the rules for entering each country. Currently these rules are
defined by placeholder rules from a YAML file which matches the current
structure of the new API entries. The intention is to replace the YAML
file with a call to the world location API once it has been updated with
the COVID rules.