Display travel rules for world destination
Display travel rules for world destination on results page of the
COVID-19 travel smart-answer.

These are links to the rules for entering a world location. The
guidance for entering England will be added in later commits.

A hard-coded list is used to determine whether to show deep links
to the foreign travel advice page. Hopefully there should come a
time when all of the travel advice pages follow the same heading
structure and this list will no longer be required.