Add new outcome to the business_coronavirus_support_finder flow
Add the new Omicron Hospitality and Leisure Grant that has been made
available to the business_coronavirus_support_finder flow.

This allows businesses that have been negatively affected by the
Omicron variant to access financial support.

The scheme needs to be displayed if the folliwing rules are true:

* The business is based in 'England'
* The type of business is 'Retail, hospitality or leisure'

For all other cases it is not shown.

When applicable, it should be shown as the first result in the
'Coronavirus schemes, grants and loans' section on the results page.