Get more covid status information from the data
Gets more covid status information from the World Location API.

It uses the safe navigation operator (`&`) and `digs` for the data to
prevent `NilClass` exceptions as  there is no guarantee that every world
location will have COVID status data.

At the moment a YAML file is being used instead of the actually API as
whilst the COVID status information is being added to the API, but the
YAML will be replaced with an API call in the future.