Get the next COVID status based on the date
Gets all of the COVID statuses for a world location that are in the
future and finds the closest to the current time, or the "min" date to
find the most recent. This uses the same logic that was used in the
Local Restrictions postcode checker [1] to find the most recent status.

It uses the safe navigation operator (`&`) and `digs` for the data to
prevent `NilClass` exceptions as  there is no guarantee that every world
location will have COVID status data.

[1]: https://github.com/alphagov/collections/commit/98851bbfaa511987617eaf42fac79ecaca3a60b5