Return nil if no COVID status exists
As the data is being manually updated in a YAML file it's possible that
COVID status data will not exist for a world location. In those cases we
want to gracefully return nil, so that anything that checks for a
particular status, i.e. if status is "red", will return false.