Add inset text content before country results
This content applies to all of the countries listed below it.

Adds a new class to give the left column a height of 1px. We can't use
0px, as that is the equivalent of not having a column and makes the
inset text full width. Adding a class, rather than continuing to use
`<div class="govuk-grid-column-one-third">&nbsp;</div>` means that
the columns stack correctly in mobile view and don't have a gap between
sections.

Updates the other views to use the new class to be consistent and to
remove the extra line (column) spaces in mobile view.