Add meta tag configuration for vaccination_status
Adds a meta tag to the layout that enables the publishing component gem to strip vaccination status out of the URL. This will not work without the line linked below in the publishing components gem. https://github.com/alphagov/govuk_publishing_components/pull/2582/files#diff-922817ed31d14664e3e1ea8c60eebd79ca15a13f7bbf35b690da73b0ee202188R29

This works because publishing components has to check for the exstence of the tag created in this commit.

This is being done as part of a wider piece of work to remove vaccination status from out logs because this is considered PII/ personal health data