Add test to check that vaccination_status has not been renamed
This test exists so that if the name of the `vaccination_status` question is ever renamed the test suite fails. This needs to happen because if that question name changes without making other changes we could start exposing vaccination status to GA. Vaccination status is deemed personal health informaion and it could land us in trouble if we start exposing/collecting it.

Generally speaking URLS are sent to GA and this is how the overwhelming majority of applications expose data to GA. Smart answers is no exception to this and this PR removes the response to the `vaccination_status` question from the URL before it's sent to Google Analyitcs https://github.com/alphagov/govuk_publishing_components/pull/2582. If the title of the `vaccination_status` question is changed the regex in that PR will start to silently fail and expose `vaccination_status` to GA through the URL. If the `vaccination_status` question title is changed it must be changed here as well to avoid exposing user health information https://github.com/alphagov/govuk_publishing_components/pull/2582/files#diff-922817ed31d14664e3e1ea8c60eebd79ca15a13f7bbf35b690da73b0ee202188R16

Every time a user completes a question on Smart Answers the response is tracked and sent to GA in plain text. This is in addion to the behaviour described above that sends the URL to GA. These things work separately and require changes to be made both in smart answers and the publishing component gem. https://github.com/alphagov/smart-answers/pull/5725 alters this behaviour for the `vaccination_status` question to redact the response from the key value pair. This ensures that anyone looking at the raw responses in GA will be stopped from seeing any responses to this question because we never save them in the first place. Changing the title of the question in smart answers would cause thiss fuctionality to fail because `vaccination_status` would no longer exist and so couldn't be filtered and the plain text would be exposed. If the question title has to be changed it must be changed here too https://github.com/alphagov/smart-answers/pull/5725/files#diff-4afbf46f26483a518d48d1d8d77371e13d8f7fe0620741ecedae1335036ba4e2R14.