Logic changes for travelling with young people
It's been asked that there's different behaviour depending on whether the user is travelling to a red list country or not. This makes it so that if the user is travelling to a red list country then nothing changes, but if they're not they're taken to a new question about children that only cares if they're below 17.