Add travel helper methods
Different content will be displayed if the user is travelling Ireland in
future commits.

This adds methods to help determine which content to display.

Implemented as separate methods as we will need to know if the user has
only travelled to Ireland or if they have travelled to other countries
as well so that different content can be rendered.