Codify vaccination_status in check_travel_during_coronavirus flow
Currently the `vaccination_status` question in the `check_travel_during_coronavirus` flow has the following possible responses:

* "vaccinated": "I will have had a second dose of an approved COVID-19 vaccine 14 or more days before I travel"
* "in_trial": "I'm taking part in an approved COVID-19 vaccine trial in the UK"
* "exempt": "I cannot have a COVID-19 vaccination for a medical reason approved by a clinician"
* "none": "None of the above"

As this flow is URL-based, we would see something like this `vaccination_status=vaccinated` in the URL when the question is answered.

We need to obfuscate (codify) the response in the URL. For example:

* "123456": "I will have had a second dose of an approved COVID-19 vaccine 14 or more days before I travel"
* "234567": "I'm taking part in an approved COVID-19 vaccine trial in the UK"
* "345678": "I cannot have a COVID-19 vaccination for a medical reason approved by a clinician"
* "456789": "None of the above"

Vaccination status is considered PII Health data and so we need to replace this question/response as plain text from the URL with a codified version.