Fix broken test in check_travel_during_coronavirus flow
Page does not have the text `people who aren't fully vaccinated` yet the test
was passing!

Update test (and the converse following one) to use the header text instead
of text that is expected somewhere on the page, as this is much less likely
to change and will break more tests when/if it does.