Add header between Ireland and rest of content
If your are travelling to Ireland as well as another country, extra
content appears. However if you are fully vaccinated, it's not clear
where the Ireland content ends and the more generic content begins, so
the content in the H2 header has been duplicated.

The "not vaccinated" content does not require a similar header because
it already has a "Before you travel to England" header that provides the
same separation.