Add a new helper method travelling_with_children
Travelling with children is only asked if the user is travelling to a
red list country, so we need to check if the question has been answered
at all, before making sure "none" was not selected.

This method will be used in the next commit to conditionally show a deep
link to guidance for travelling with young people.