Get radio button vaccination options from config
The vaccination statuses were changed from a readable value to an
obscure code in https://github.com/alphagov/smart-answers/pull/5764 to
protect users privacy as vaccination status is considered PII Health
data.

Whilst great for protecting users, codifying the vaccination status
makes it easier to get the statuses confused when the source code is
updated.

The aim of this change is to make it easier to update the smart-answer
flow by storing the possible vaccination statuses in config and doing
a lookup to populate the radio button questions, evaluate logic in the
outcomes and in the tests.

The first step is updating the radio button to render it's content from
the config. The controller and view logic will be updated in the
following commits.