Add validation to transiting question
Validate that users cannot select all of their previously selected
countries as transiting countries. At least one country must be
unchecked to be considered a destination country.

In the results, the guidance for transiting countries is much less than
the guidance for destination countries. This check guards against users
misreading the question, selecting all of the countries thinking they're
selecting destinations and only being provided with limited guidance.

The error message is displayed using a custom `error_no_destination`
block, so that the default "Please answer this question" error message
will be displayed if the user doesn't select anything.