Config changes for new i18n setup
In order to make the locale file changes work I've had to add a line to the application.rb to change the load path for i18n. The default path that rails looks for is `config/locales/en.yml`. The `locales` folder can have multiple files in it that cover different languages, en for english, fr for french etc. This approach works well in normal applications where all of the internationalised content can live in the same place without creating a vast file. It wouldn't work as well for smart answers because it would meant that all the internationalised content for all the smart answers would be in one file. If there were ever a move to internationalise smart answers this file would be thousands upon thousands of lines long and very difficult to navigate and work with. What I've done here is to add a folder with the same name as the smart answer it holds internationalised content for. In this case that means `config/locales/check_travel_during_coronavirus/en.yml`. This is in order that an individual `en.yml` file only holds content for one smart answer, thus rendering that file easier to deal with should more content be internationalised. The changes I've made to the `application.rb` file reflect this change and let rails know that there's a new default place to look for internationalised content. This in turn means that this content can be rendered in the views without any unusual paths. I've left the original en.yml file so it can be used for shared content if the approach to i18n is ever changed for smart answers.