Add HTTParty gem
What

This change adds the HTTParty gem to the application.

Why

It was discovered that a similar HTTP client - Faraday - is a dependency of the following gems:

* sentry-ruby
* sentry-ruby-core
* zendesk_api

Which themseleves are also implicit dependencies.

Faraday was initially used, however appeared to be quite slow in use. So, we decided
to benchmark some popular Ruby HTTP clients.

Here are the results conducted on a macOS Catalina 10.15.7, 2.6 GHz 6-Core Intel Core i7,
with 16 GB 2667 MHz DDR4.

The clients tested are:

* Faraday:     https://github.com/lostisland/faraday
* HTTParty:    https://github.com/jnunemaker/httparty
* Rest-Client: https://github.com/rest-client/rest-client
* HTTP:        https://github.com/httprb/http

The results...

6 iterations each of: https://www.gov.uk/api/content/foreign-travel-advice
                 user     system      total          real
Faraday:     0.107038   0.011063   0.118101   (  0.691918)
HTTParty:    0.053854   0.006775   0.060629   (  0.499783)
Rest-Client: 0.130284   0.011241   0.141525   (  0.573892)
HTTP:        0.048266   0.010650   0.058916   (  1.079538)

6 iterations each of: https://www.gov.uk/api/content/foreign-travel-advice/spain
                 user     system      total          real
Faraday:     0.121543   0.012714   0.134257   (  0.701960)
HTTParty:    0.034293   0.007447   0.041740   (  0.572652)
Rest-Client: 0.098948   0.008438   0.107386   (  0.599544)
HTTP:        0.026310   0.008052   0.034362   (  0.885937)

Based on the above benchmarks it was decided to go with HTTParty instead
of sticking with Faraday for performance reasons.