Add foreign travel advice header checker
What

This change adds a service that checks if specific countries have had their
headings converted to the agreed format by the FCDO content design team.

What does all that mean?

FCDO maintain foreign travel advice content for each country which is
published and made available at https://www.gov.uk/foreign-travel-advice.
For example: Spain (https://www.gov.uk/foreign-travel-advice/spain). They
create and maintain this content in Travel Advice Publisher
(https://travel-advice-publisher.publishing.service.gov.uk/) where each country
has a summary content block and a number of sections (called 'parts' in the
app).

FCDO only want to maintain this content once - in one system.

Each 'part' has a 'title', 'body' and a 'slug'. The 'body' is markdown enabled,
so is effectively an arbitrarily sized chunk of HTML - some are extremely
large! Each 'part' 'body' then becomes a sub-page under the main country page
(e.g. https://www.gov.uk/foreign-travel-advice/spain/entry-requirements).

It is worth noting that the editor has absolute control of the content within
a 'part' 'body', 'title' and 'slug'.

As stated the 'body' itself is markdown, so the content can be sectioned further
by introducing headers - for example: '##All travellers', which gets turned
into HTML like this '<h2 id="all-travellers">All travellers</h2>'.

The agreed format with FCDO - and what "converting headers" means - is that
FCDO have agreed to create and maintain the following headers within the 'Entry
requirements' 'part' for each - and every - country:

* All travellers ("all-travellers")
* If you're transiting through [COUNTRY NAME] ("if-youre-transiting-through-COUNTRY_NAME")
* If you're not fully vaccinated ("if-youre-not-fully-vaccinated")
* If you're fully vaccinated ("if-youre-fully-vaccinated")
* Children and young people ("children-and-young-people")
* Exemptions ("exemptions")

As the id used in the header is auto-generated when the content is published,
any change to the header itself also changes the id, and therefore the link to
that content. A key requirement of the results that are presented back to the
user by the 'check_travel_during_coronavirus' flow is to (deep) link to these
sections on the rendered page.

Why

We currently have a hard-coded list of countries that have been converted by
FCDO. This is both out-of-date and manually maintained.

The process outlined above creates a situation that is constantly at risk of
human error and whim. The slightest change to the header name means the (deep)
links on the results page no longer work. The link (when clicked) will then
fall-back to the parent 'Entry requirements' section. If the user chooses not
read the entire section - they will not receive the correct information.

When a user is completing the flow, we do not know if these sections exist (or
have been removed, renamed or moved) for any or all of the countries they are
travelling to (or through).

To mitigate this issue, we need to test if each of these sections exist or not.
So, for each of the countries that the user indicates they are going to (or
through) - we should either show (deep) links to the above sections directly if
they exist, or display a generic link to the 'Entry requirements' page itself
if they do not.