Cache individual country header statuses
Having added a service to test if a particular country has had it's
headers converted, we now want to use that in the
'check_travel_during_coronavirus' flow.

Checking these statuses each time a user completes the flow and
reaches the results page will introduce a performance hit. So, we
cache each country's header status.

Doing this means that frequently selected countries have a near
zero lookup time, and only the countries that are selected need
to be cached.

We invalidate the cache every 4 hours.