Rename user research banner component
In the check-travel-during-coronavirus smart answer we want to reuse this
banner component.

The banner will be used to link users to a feedback survey, this differs
from the current banner which is used to link users to become research
participants. The rename allows us to use the banner component in
both contexts.