Render user intent banner on travel c19 results
The user intent banner has been rendered on the
check-travel-during-coronavirus results page in order to gather extra
feedback. This is because the tool has been launched recently so we want
to encourage users to give feedback on it.

The user intent banner is also used on the /next-steps-for-your-business flow across all pages.
 It is rendered within the application layout and applied only to the
/next-steps-for-your-business flow using a conditional.

We investigated refactoring this, however it was complex to implement
for the various presenters, especially as it is only used for the
next-steps-for-your-business flow. Given that we have a custom layout
for the check-travel-during-coronavirus flow and we only need to display
the banner on the results page, we have added our implementation into there
rather than attempting a full refactor. This is partly due to time constraints
of wanting to get this banner out ASAP and also due to not knowing if the
next-steps-for-your-business user research is still utilised for research
(maybe it could be removed). We expect the banner on the check-travel-during-coronavirus
flow to be relatively temporary.