Add a "redact" property to the question
This takes the initial concept of using a method in the question_presenter
to determine the label shown, but rather than hardcoding the vaccination_status
question, a property is set on the question in a similar manner to the `template_name`
property.

The new property is defined on the "node". I'm not sure if is should be there on
in "question" which inherits from node. However with all of the inheritance, in smart-answers,
putting the property in "node" made it easier to add a "redacted?" method to
NodePresenter for use in QuestionPresenter.

The original `change_link_track_label` test only needs to stub the value of `redacted?`
to pass.