Functionality to set 'redact' flag
The response to vaccination_status is health data.
Despite there not being any personally identifing questions in this Smart Answer,
we are required to prevent health data from being sent to Google Analytics.

Currently the response is sent as part of the 'change' link event tracking label. Adding
the `redact` method allows this value to be set to true for individual questions. This value
will be used in a future commit to set the tracking label text. If other question responses
 need to be redacted in future then this flag can simply be applied.

A cop was disabled as Rubo was complaining about `def redact(filter = false)`. Changing this to `def redact(filter: false)` meant that
`redact true` has to change to `redact filter: true` in the flow which doesn't look nice. Someone else had a similiar problem
[here](https://github.com/alphagov/smart-answers/blob/main/lib/smart_answer/erb_renderer/question_options_helper.rb#L7) so we have
taken inspiration and disabled it.

Co-authored-by: Leena Gupte <leena.gupte@digital.cabinet-office.gov.uk> as a result of approach suggested on [this branch](991964c)