Add methods to handle multiple countries that don't require plf
This change adds the following to the check_travel_during_coronavirus flow
calculator:

* A static list of countries that do not require a Passenger Locator Form (PLF)
to be filled in when returning to England from.

* Methods to identify which - if any - PLF exempt countries the user is returning
from and if they are returning from any at all.

* A method that identifies if the user is returning from a visit to a single PLF
exempt country.

* A method to build a human readable list of countries from a given list of countries.
This returns a formatted string such as "South Africa", "South Africa or Spain", or
"South Africa, Spain or USA".

* Two methods that use the human readable list of countries method - one for the
PLF exempt countries that the user is going to, and one for all PLF exempt countries
(this is used on the start page - before we are given the actual countries that the
user is travelling to)

Later commits will make use of these methods.

Co-authored-by: Jon <jonathan.hallam@digital.cabinet-office.gov.uk>