Update links on the generic outcome page
We currently say 'lawyers and translators/interpreters' but only provide a link for lawyers. Including a separate link for translators will be helpful for users.

Replacing the individual country links with generic ones for all countries will make this smart answer easier to maintain.

Regions have been removed, they were only previously used for Cyprus. After checking the previous links for Cyprus, you would be redirected to a single page that includes content for both regions.

The updated links will still only be displayed for a country if the "lawyer" key is present in the prisoner_packs.yml file.