Fix to stop duplicate GA events firing
Update `tracks-results.js` to find all links within the element that the module is acting on, instead of all of the document.

Ensure custom data tracking attributes included on any link within the module are used instead of the default values. Added a test to check this happens.

The new `div` element added to the wrap both the `body` and `next_steps` sections of the results page templates, ensures that links in both sections are tracked as before.

Removed the `gem-track-click module` from the `result-card` component. This approach should not be used in a component as it makes it less flexible.