Add uprated_value_of_property method to calculator
This method reads from a YAML file listing years with associated
uprating values. The value_of_property is multiplied
by the rating for the year_of_purchase to give an uprated value,
accounting for the current estimated value of the property.

If the year_of_purchase is not present (currently, before 1970) a
default uprating value is used.