Raise specific exception for attempting an invalid transition
This change has been put in to allow fixing a behavioural regression
introduced by 152f74a. That change worked on the assumption that
InvalidNode was only used for cases where a node was misconfigured.

It transpires that this exception was also used as a means of handling a
transition to a node that is impossible: trying to reach a subsequent
node after reaching an outcome.

As this does not represent a configuration issue and instead a problem
with user input, this should not be handled with an InvalidNode
exception. To handle this I've created a new exception
InvalidTransition.