Return a 404 response on an invalid transition.
If a user tries to provide more responses to a Smart Answer after
they've reached an outcome the application will raise an error. This PR
updates the error handling so that this error, InvalidTransition, will
return a 404 response.

An example page that demonstrates this is: https://www.gov.uk/state-pension-age/y/age/2022-01-01/male

I've added a unit test to capture this scenario for a
SmartAnswerController.

As far as I can tell, it is not possible to get this exception when
working with the FlowController because of this logic [1] which pauses
execution at an outcome. Therefore I have not added a test for this.

However as a safety measure (and duplication reduction) I've moved the
exception handling into the ApplicationController.

[1]: https://github.com/alphagov/smart-answers/blob/86f5bc78e240600c3a83a47bccee8f50f3a384db/lib/smart_answer/state_resolver.rb#L46-L50