Add YearQuestion type
Adds a new question type of Year. A user can submit
a year which is passed in as a hash in the format
`year: YYYY` when the question is initially answered.

When rendering subsequent questions/nodes, the year is
instead passed as a string.

In both cases, we ensure that the passed input can be
used to create a valid Date object.

Ultimately parsed input is returned as submitted, having
been successfully processed.