Add Sure Start Maternity Grant outcome
As part of this we've added the conditional which I've abstracted into a
general child eligibility checker, as there are quite a few result
outcomes to be added that have pretty much the same logic so can by DRYd
up.

We'll only have one set of tests for these group of outcomes, as the
only difference will be the lists `skip_benefit_list` and
`eligible_child_ages` and this falls under config, so doesn't need to be
tested.

Each outcome will also have a flow test so there will be no drop in test
coverage.