Add radio with intro question
This pattern is supported by the design system [1] and has been
requested for the Cost of Living checker V2 Smart Answer.

The radio component [2] does not support this format and so we need to
update Smart Answers. My original thinking was to update the existing
radio question. However, Smart Answer questions are not really designed
to support variances in format and so a new question seemed the cleanest
solution.

[1]: https://design-system.service.gov.uk/patterns/question-pages/#asking-complex-questions-without-using-hint-text
[2]: https://github.com/alphagov/govuk_publishing_components/blob/main/app/views/govuk_publishing_components/components/_radio.html.erb