Update outcomes for marriage abroad Malaysia
As a separate request, 'Where do you live?' question is being removed
from this smart answer for Malaysia. New headings added 'Prove you're
free to get married', 'If you or your partner is a Muslin' and
subheadings within those sections.  Details added so users know what
they need to take to the appointment and who needs to attend based on
the outcome.