Remove Qatar from EVW list and fix broken flow
There was no condition for ETA countries within the check_uk_vis_flow.rb because Qatar was still in the EVW list and so we were able to just use the EVW condition to ensure the correct outcome is shown. However, now that Qatar has been removed entirely from the EVW list, an ETA condition is required to ensure the flow still leads to the correct outcomes.