Error link in Radio w/ Intro pages now work
When the radio_with_intro page type was added, the array of types in the
"error_id" method of the question_presenter was not updated. This meant
that errors for the radio_with_intro pages would not link to the where
the error had taken place. This has been fixed and a test has been
added.