Simplify `per_hour_minimum_wage`
This method in the `SmartAnswer::Calculators::MinimumWageCalculator` class had a `date` parameter that was never passed by callers and so always used its default value.

The call to `rates_for_date` is also unnecessary; the data from that method call is already stored in `@minimum_wage_data` so simply use that field directly.