Update uk_result.erb
Trello: https://trello.com/c/5ZA9blV6/4384-9-sept-changes-to-the-death-certification-process

Summary:

From 9 September 2024, a medical examiner must independently review the proposed cause of death on the medical certificate of cause of death (MCCD) of all deaths, except those that need to be referred to a coroner.

Only once the registrar has received the MCCD from the medical examiner, and the medical examiner has called the user to let them know that’s happened, can users register a death. It must be done within 5 days. 

In addition to this, users will no longer need to complete an ‘application for cremation’, instead they’ll get a certificate for cremation. 

We have therefore made it clear that users need to wait for a call from the medical examiner in order to register a death.