Update countries to require an ETA
Countries that were in the `COUNTRY_GROUP_ETA_ROLLOUT_GROUP_1_REST_OF_THE_WORLD` group should now be requiring an electronic travel authorisation (ETA). All countries have been moved, except for Taiwan, which has some special handling rules to be grappled with in a future commit.

This change has caused a number of flow changes and so tests have been updated to reflect this.

Hong Kong and Macao are the only countries that can get to the `:what_sort_of_travel_document?` question (as seen in the next_node block of Q1). Both of these countries now require an ETA. This makes it impossible to get to the:
- `:outcome_study_m`
- `:outcome_tourism_n`
- `travelling_visiting_partner_family_member?`
- `:outcome_medical_y`
- `:outcome_transit_to_the_republic_of_ireland`
- `:outcome_work_m`
nodes for these countries (referred to as travel document countries (see `@travel_document_country`) in the tests). It also means that when travelling to the Republic of Ireland, travellers will be directed to the `:outcome_transit_to_the_republic_of_ireland_electronic_travel_authorisation` outcome instead of the `:outcome_no_visa_needed` outcome.

Taiwan is the only country left in the “ETA rollout group 1” list, and it has its own outcomes, meaning you can no longer get to the:
- `:outcome_tourism_n`
- `:outcome_medical_n`
- `:outcome_study_no_visa_needed`
- `:outcome_marriage_nvn`
- `:outcome_joining_family_nvn`
outcomes for “ETA rollout group 1” countries—those tests have been deleted.

British national overseas passport holders now require an ETA, so will not be shown the text saying that they do not need a visa for the `:outcome_marriage_nvn` outcome.